requirements.txt
opencv-python
numpy
Pillow
face_recognition
customtkinter
mysql-connector-python
CTkMessagebox
CTkTable
CTkListbox
pandas
xlsxwriter
openpyxl
pyinstaller